docker-images/app/requirements.txt
entsoe-py==0.6.8
fastapi==0.112.2
joblib==1.4.2
lightgbm==4.5.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.5.2
tqdm==4.66.5